microservices/core/pyproject.toml
[tool.poetry]
name = "epomatti_aksmrc_core"
version = "0.1.0"
description = ""
authors = ["Evandro Pomatti <evandro.pomatti@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pymongo = "^4.1.1"
azure-identity = "^1.10.0"
azure-keyvault-secrets = "^4.4.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
pylint = "^2.13.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
